modified type of labels in the input_functions

+++ main/train_using_tf_estimator.py
@@ -29,7 +29,7 @@ def train(labels_file, data_dir, desired_image_size, architecture, num_hidden_un
     train_input_fn = input_fn(
         x_feed_dict={"x": np.array(x_train),
                      "seq_lens": dataset_utils.get_seq_lens(x_train)},
-        y=np.array(y_train),
+        y=np.array(y_train, dtype=np.int32),
         num_epochs=num_epochs,
         batch_size=batch_size
     )
@@ -37,7 +37,7 @@ def train(labels_file, data_dir, desired_image_size, architecture, num_hidden_un
     validation_input_fn = input_fn(
         x_feed_dict={"x": np.array(x_test),
                      "seq_lens": dataset_utils.get_seq_lens(x_test)},
-        y=np.array(y_test),
+        y=np.array(y_test, dtype=np.int32),
         shuffle=False
     )
